+++ Editor/AssetChange.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
  * Copyright (c) 2022 Willy Alberto Kuster
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -19,26 +19,26 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-using UnityEditor;
-using UnityEngine;
-
 namespace Willykc.Templ.Editor
 {
-    [TemplEntryInfo(ChangeType.Import)]
-    internal sealed class ScriptableObjectEntry : TemplEntry
+    public struct AssetChange
     {
-        [TemplInput]
-        public ScriptableObject scriptableObject;
+        private const string Empty = "";
 
-        protected override bool IsValidInputField =>
-            scriptableObject &&
-            !(scriptableObject is ScribanAsset) &&
-            !(scriptableObject is TemplSettings);
+        public ChangeType type;
+        public string currentPath;
+        public string previousPath;
 
-        protected override object InputValue => scriptableObject;
-
-        protected override bool IsInputChanged(AssetChange change) =>
-            change.currentPath == AssetDatabase.GetAssetPath(scriptableObject);
-
+        internal AssetChange(
+            ChangeType changeType,
+            string currentPath,
+            string previousPath = Empty)
+        {
+            this.type = changeType;
+            this.currentPath = currentPath
+                ?? throw new System.ArgumentNullException(nameof(currentPath));
+            this.previousPath = previousPath
+                ?? throw new System.ArgumentNullException(nameof(previousPath));
+        }
     }
 }