+++ Samples~/Extensions/TemplFunctions.cs
@@ -1,7 +1,0 @@
-namespace Willykc.Templ.Editor
-{
-    public static class TemplFunctions
-    {
-        public static string Padding(string str) => $"_{str}_";
-    }
-}