+++ Samples~/Extensions/TemplFunctionsSample.cs
@@ -1,0 +1,10 @@
+namespace Willykc.Templ.Samples
+{
+    using Editor;
+
+    [TemplFunctions]
+    public static class TemplFunctionsSample
+    {
+        public static string Padding(string str) => $"_{str}_";
+    }
+}